+++ .dockerignore
@@ -2,6 +2,7 @@ node_modules
 .git
 dist
 Dockerfile
+.env
 
 # OS
 .DS_Store
