+++ .env.example
@@ -1,7 +1,7 @@
 PORT=4000
 
 DB_PORT=5432
-DB_HOST=db
+DB_HOST=localhost
 DB_USER=admin
 DB_PASSWORD=admin
 DB_NAME=hls_db
