+++ docker-compose.yml
@@ -7,6 +7,8 @@ services:
     container_name: hls-app
     restart: always
     env_file: .env
+    environment:
+      - DB_HOST=db
     ports:
       - ${PORT}:${PORT}
     depends_on:
