+++ src/api/album/entities/album.entity.ts
