+++ src/api/album/interfaces/album.interface.ts
