+++ src/api/artist/artist.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { EntitiesList } from '../utils/constants';
-import { EntityNotFoundError } from '../errors';
+import { EntitiesList } from '../../utils/constants';
+import { EntityNotFoundError } from '../../errors';
 import { CreateArtistDto } from './dto/create-artist.dto';
 import { UpdateArtistDto } from './dto/update-artist.dto';
 import { ArtistEntity } from './entities/artist.entity';
