+++ src/api/artist/dto/create-artist.dto.ts
