+++ src/api/artist/dto/update-artist.dto.ts
