+++ src/api/artist/entities/artist.entity.ts
