+++ src/api/artist/interfaces/artist.interface.ts
