+++ src/api/favorites/entities/favAlbum.entity.ts
