+++ src/api/favorites/entities/favArtist.entity.ts
