+++ src/api/favorites/interfaces/favorites.interface.ts
