+++ src/api/user/user.service.ts
@@ -2,8 +2,8 @@ import { Injectable, ForbiddenException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm/dist';
 import { Repository } from 'typeorm';
 import * as bcrypt from 'bcrypt';
-import { EntityConflictError, EntityNotFoundError } from '../errors';
-import { EntitiesList } from '../utils/constants';
+import { EntityConflictError, EntityNotFoundError } from '../../errors';
+import { EntitiesList } from '../../utils/constants';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdatePasswordDto } from './dto/update-password.dto';
 import { UserEntity } from './entities/user.entity';
